add sound for hitting doorway
“ I didn't want players to have trouble going through a doorway.
So when you approach a doorway, the game does the opposite of what happens
if you push an object - so essentially the doorway would push the player
into the centre, and through into the next room. I played games where
you had to go backwards and forwards, lining up, and I felt it much better
not to have to fuss over things like that. ”
~~ Jon Ritman

and update graphics of helicopter-bug & book in present set

and refresh scripts for OS X